pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "electoral-sim"
version = "0.1.0"
description = "A spatial voting simulator for comparing electoral systems"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "pyyaml>=6.0",
    "gymnasium>=0.29",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pandas>=2.0",
    "pytest>=7.0",
]

[project.optional-dependencies]
dev = ["jupyter", "ipykernel", "black", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
include = ["electoral_sim*"]
